densenet_function/requirements.txt
onnxruntime==1.15.1
pillow==10.1.0
numpy==1.24.4
boto3==1.28.77
fastapi==0.104.1
mangum==0.17.0
uvicorn[standard]==0.24.0
pydantic==2.4.2
docker==6.1.3